feat: move library to use native big int

+++ src/types/hash.ts
@@ -1,5 +1,4 @@
 import { Bytes } from './index';
-import { BigInteger } from 'big-integer';
 import Hash from '../lib/hash/hash';
 
 export interface IHash {
@@ -8,5 +7,5 @@ export interface IHash {
   String: () => string;
   Hex: () => string;
   Equals: (hash: Hash) => boolean;
-  BigInt(): BigInteger;
+  BigInt(): bigint;
 }
